RabbitMQ/requirements.txt
grpcio==1.70.0
grpcio-tools==1.70.0
requests==2.32.3
protobuf==5.29.3
pika==1.3.2